backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.13.1

# Auth
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
argon2-cffi>=21.3.0

# Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# AkitaLLM - Motor de orquestração
akitallm>=1.0.0

# Utils
python-dotenv>=1.0.0
httpx>=0.26.0

# Dev/Test
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
